add testing for delete comment endpoint

+++ db/models/comments.model.js
@@ -28,18 +28,15 @@ exports.addComment = (comment, username, id) => {
 };
 
 exports.removeComment = (id) => {
+  if (isNaN(id)) return Promise.reject({ status: 400, msg: "bad request" });
   return db
     .query("DELETE FROM comments WHERE comment_id = $1 RETURNING *", [id])
-    .then((res) => {
-      return { comment: res.rows[0] };
+    .then(({ rows }) => {
+      if (rows.length === 0)
+        return Promise.reject({
+          status: 404,
+          msg: "no comment matching that id",
+        });
+      return { comment: rows[0] };
     });
 };
-
-// {
-//     comment_id: 1,
-//     body: "Oh, I've got compassion running out of my nose, pal! I'm the Sultan of Sentiment!",
-//     article_id: 9,
-//     author: 'butter_bridge',
-//     votes: 16,
-//     created_at: 2020-04-06T12:17:00.000Z
-//   },
